feat: add proposed changes
